collection: Only add bear base directory to path
The idea is to remove the `sys.path` modification completely, though
this is not completely possible with our infrastructure. There's an
issue on Windows, where the unpickling of modules is performed by the
qualifier name (i.e. like you would import python modules) and not by
the actual module path. This is needed by our multiprocessing machinery,
so at least the bear base directory only is appended to `sys.path` and
a qualifier name is determined from coala so imports work smoothly in
our core.